python/requirements.txt
librosa
numpy
pyzmq
websockets